requirements.txt
google-generativeai
requests
python-dotenv
fastapi
uvicorn
pydantic
psutil